Tutor fix, add creeper and noobic

+++ Assets/_PROJECT/Scripts/GachaPets/Configs/PetStationConfig.cs
@@ -3,5 +3,5 @@
 [CreateAssetMenu(fileName = "PetStationConfig", menuName = "Configs/PetStationConfig")]
 public class PetStationConfig : ScriptableObject {
     [field: SerializeField] public PetChance[] Pets { get; private set; }
-    [field: SerializeField] public float Price { get; private set; }
-}
+    [field: SerializeField] public long Price { get; private set; }
+}
